+++ functions.php
@@ -5,7 +5,7 @@
  * @package Teznevise
  */
 if ( ! defined( 'ABSPATH' ) ) { exit; }
-define( 'TEZNEVISE_VERSION', '1.9.1' );
+define( 'TEZNEVISE_VERSION', '1.9.2' );
 define( 'TEZNEVISE_DIR', get_template_directory() );
 define( 'TEZNEVISE_URI', get_template_directory_uri() );
 
@@ -110,6 +110,9 @@ function teznevise_enqueue_assets() {
 	wp_enqueue_style( 'teznevise-pages', TEZNEVISE_URI . '/assets/css/pages.css', array( 'teznevise-components' ), TEZNEVISE_VERSION );
 	wp_enqueue_style( 'teznevise-chrome', TEZNEVISE_URI . '/assets/css/chrome.css', array( 'teznevise-pages' ), TEZNEVISE_VERSION );
 	wp_enqueue_style( 'teznevise-modernization', TEZNEVISE_URI . '/assets/css/modernization.css', array( 'teznevise-chrome' ), TEZNEVISE_VERSION );
+	wp_enqueue_style( 'teznevise-legacy-wpcode', TEZNEVISE_URI . '/assets/css/legacy-wpcode.css', array( 'teznevise-modernization' ), TEZNEVISE_VERSION );
+	wp_enqueue_script( 'teznevise-calculators', TEZNEVISE_URI . '/assets/js/calculators.js', array(), TEZNEVISE_VERSION, true );
+	wp_script_add_data( 'teznevise-calculators', 'strategy', 'defer' );
 
 	wp_enqueue_script( 'teznevise-chrome', TEZNEVISE_URI . '/assets/js/chrome.js', array(), TEZNEVISE_VERSION, true );
 	wp_script_add_data( 'teznevise-chrome', 'strategy', 'defer' );